+++ code/ylikuutio/time/time.cpp
@@ -23,8 +23,8 @@
 
 namespace yli::time
 {
-    float get_time()
+    double get_time()
     {
-        return static_cast<float>(SDL_GetTicks()) / 1000; // `SDL_GetTicks` return milliseconds.
+        return SDL_GetTicks() / 1000; // `SDL_GetTicks` return milliseconds.
     }
 }
